requirements.txt
opencv_python==4.5.1.48
Flask==2.0.1
pan_aadhar_ocr==1.0.4
pdf2image==1.16.0
pytesseract==0.3.7
Pillow==8.3.1
